Refactoring classe helper, implementação do método main
