+++ src/main/java/br/com/viagemfundomar/helper/ViagemAoFundoDoMarHelper.java
@@ -1,41 +1,65 @@
 package br.com.viagemfundomar.helper;
 
 import br.com.viagemfundomar.enumarator.Direcao;
-import br.com.viagemfundomar.enumarator.Instrucao;
 import br.com.viagemfundomar.model.Submarino;
 
 public class ViagemAoFundoDoMarHelper {
 
+	private static final char DOWN = 'D';
+	private static final char UP = 'U';
+	private static final char LEFT = 'L';
+	private static final char RIGHT = 'R';
+	private static final char MOVE = 'M';
+	
 	public void calcularInstrucoes(char[] coordenadas, Submarino submarino){
-		for (char c : coordenadas) {
-			if(Instrucao.DOWN.getAbreviacao() == c){
+		for (char coordenada : coordenadas) {
+			switch (coordenada) {
+			case DOWN:
 				submarino.setZ(submarino.getZ()-1);
-			} else if(Instrucao.UP.getAbreviacao() == c){
+				break;
+
+			case UP:
 				submarino.setZ(submarino.getZ()+1);
-			} else if(Instrucao.LEFT.getAbreviacao() == c){
+				break;
+				
+			case LEFT:
 				if(submarino.getDirecao() == 1){
 					submarino.setDirecao(Direcao.OESTE.getDirecao());
 				}else{
 					submarino.setDirecao(submarino.getDirecao() - 1);
 				}
-			} else if(Instrucao.RIGHT.getAbreviacao() == c){
+				break;
+			
+			case RIGHT:
 				if(submarino.getDirecao() == 4){
 					submarino.setDirecao(Direcao.NORTE.getDirecao());
 				}else{
 					submarino.setDirecao(submarino.getDirecao() + 1);
 				}
-			}else if(Instrucao.MOVE.getAbreviacao() == c){
-				if(submarino.getDirecao() == Direcao.OESTE.getDirecao()){
-					submarino.setX(submarino.getX()+1);
-				}else if(submarino.getDirecao() == Direcao.LESTE.getDirecao()){
-					submarino.setX(submarino.getX()+1);
-				}else if(submarino.getDirecao() == Direcao.NORTE.getDirecao()){
-					submarino.setY(submarino.getY()+1);
-				}else if(submarino.getDirecao() == Direcao.SUL.getDirecao()){
-					submarino.setY(submarino.getY()-1);
-				}
+				break;	
+				
+			case MOVE:
+				moverSubmarino(submarino);
+				break;
+				
+			default:
+				break;
 			}
-			System.out.println(submarino);
+			
+			
+			
+		}
+	}
+
+	private void moverSubmarino(Submarino submarino) {
+		if(submarino.getDirecao() == Direcao.OESTE.getDirecao()){
+			submarino.setX(submarino.getX()+1);
+		}else if(submarino.getDirecao() == Direcao.LESTE.getDirecao()){
+			submarino.setX(submarino.getX()+1);
+		}else if(submarino.getDirecao() == Direcao.NORTE.getDirecao()){
+			submarino.setY(submarino.getY()+1);
+		}else if(submarino.getDirecao() == Direcao.SUL.getDirecao()){
+			submarino.setY(submarino.getY()-1);
 		}
 	}
 }
